requirements.txt
torch>=2.0
torchvision>=0.15
Pillow>=9.0
PyYAML>=6.0
lpips>=0.1.4
